gather some statistics from db and bugfixes
statistics for rating dream mood, clarity and sleep quality are now gathered from the database. Fixed the bug that cancelling the editing of an audio entry would delete the audio

+++ app/src/main/java/com/bitflaker/lucidsourcekit/charts/DataValue.java
@@ -1,10 +1,10 @@
 package com.bitflaker.lucidsourcekit.charts;
 
 public class DataValue {
-    int val;
+    double val;
     String label;
 
-    public DataValue(int val, String label) {
+    public DataValue(double val, String label) {
         this.val = val;
         this.label = label;
     }
